Fix incorrect statement about Apollo
Apollo does a deep comparison of the variables object when deciding whether or not the query needs to be re-executed. 

I can’t find this in the documentation, but know it to be true from experience. A quick squiz at the source code seems to indicate this is where it’s implemented:

https://github.com/trojanowski/react-apollo-hooks/blob/master/src/useQuery.ts#L119

(‘objToKey’ serialises the variables object.)