feat(useQuery): Add refetchNever config option to always use cache when it is available
feat(useQuery): Add refetchNever config option to always use cache when it is available

This features allows a config option to always use cache to serve data and never hit the API if there is a cache.  This is useful when loading static files with react-query as there is no point in revalidating the cache as it will never change. refetchOnMount is not always sufficient as sometimes user activity may unmount a component and when it is remounted the developer may still want the data to be only served from cache.

No breaking changes, new unit tets added. All unit tests passing.