fix: :bug: Make sure queries which have dataUpdatedAt of 0 will be run
When initialDataUpdatedAt is a falsy numeric value (e.g. 0), it will cause it to be replaced with Date.now(). This leads to unexpected behaviour, as one would expect the query to be invalidated instantly, if staleTime is set to an amount less than the result of Date.now(). This change will check if the value is numeric, and if so, always use this numeric value.

This changes the behaviour of the plugin for those users that have specified initialDataUpdated at <= 0.