Add data structure note and another example
Added a note about a frequent question I had and see in the discord once in a while about why the data structure for useInifiniteQuery. Also added an example for updating a single value in a page's data array. Sorry if that isn't written well enough, that was my best effort but I'm not that advanced yet!