Update paginated-queries.md
`query.data` is always `undefined` at initial render; because of this the `data.hasMore` check in the "Next Page" button throws a type error when this mini-app launches(`TypeError: Cannot read property 'hasMore' of undefined`) and crashes the whole thing. I'm using javascript optional chaning to check if there is a data object before accessing the `hasMore` property - this fixes the type error and the mini-app works fine.