feat(broadcastQuery): make BroadcastQueryClient class-based and allow broadcast of invalidations
Allow invalidations of queries in other tabs even if they're not mounted in the current tab.
Change broadcastQueryClient into a class-based BroadcastQueryClient to allow for a
clean implementation.
Add a getQueryFilters util to pick only the relevant keys in query filters, preventing an error
when trying to broadcast non-serializable objects.

BREAKING CHANGE: The experimental broadcastQueryClient has been replaced by BroadcastQueryClient