Update query-functions.md
I was looking for an example of how to handle errors thrown from a query functions when using useQuery . This page was top amongst search results. The code example under the 'Handling and Throwing Errors' section does not clearly show how to handle errors. I am adding this bit of code for any one else who lands on this page looking for an answer.