feat: Add queryKey to QueryObserverResult
When using useQueries it personally makes me nervous that you have
to rely on the array order to remap the queryKeys. By passing the
queryKey back through into the QueryObserverResult, this remapping
can more easily be done by the code that does the result
processing.

This feature was suggested here:
https://github.com/tannerlinsley/react-query/discussions/2624#discussioncomment-1344889