Return Statement
I know everyone loves the one-liner code. It looks awesome!

But in this certain case, I thought newTodo is the second argument to useMutation hook which was very confusing for me when I tried to code and it didn't work. It took me a very long time to notice that small detail.

 I I think the return statement will drastically remove that confusion. `useMutation(newTodo => axios.post('/todos', newTodo))`