feat(cancellation): Provide AbortSignal to queryFn
All queryFn calls will now have an AbortSignal instance passed to them
as part of the QueryFunctionContext. In environments that support the
AbortController/AbortSignal the consumer may use this `signal` as a
way to get notified about the query being cancelled. Consumers no
longer need to add a `.cancel()` function to the result of `queryFn()`

```ts
const query = useQuery('todos', async ({ signal }) => {
  const res = await fetch('/todos', {
    method: 'get',
    // Pass the signal to your request
    signal,
  });

  const json = await res.json();
  let data = null;

  if (!signal.aborted) {
    // do expensive CPU transformation of data
    data = expensiveTransform(json);
  }

  return data;
});
```

The previous behavior for cancellation will continue to work. If a
`.cancel()` function is provided as the result of calling `queryFn()`
the AbortSignal will not be notified when the cancellation happens.
If the runtime environment does not support AbortController/AbortSignal
the old behavior will continue to work, but the `signal` provided to
the `queryFn()` will not be notified when the cancellation happens.

Ref https://github.com/tannerlinsley/react-query/discussions/1671#discussioncomment-1310752
Ref https://github.com/tannerlinsley/react-query/discussions/1671#discussioncomment-1453013

Fix https://github.com/tannerlinsley/react-query/issues/2775
Fix https://github.com/tannerlinsley/react-query/issues/1893
Fix https://github.com/tannerlinsley/react-query/issues/1265
Fix https://github.com/tannerlinsley/react-query/issues/886