feat: better query filters
use a tri-state enum (active / inactive / all) to indicate which queries to match. It usually doesn't make sense to match no queries ("none"); However, for `refetch` on `invalidateQueries`, we can say we want to invalidate some queries, but refetch none of them. Since the default is to refetch all active queries, we need a way to override that. So `none` is allowed there