moves asides to be in context with attached bullets
I don't see how to run the docs site locally to test this, but at least on GitHub's markdown parser this would really help align the quote block with the associated bullet (as it is today is sorta hard to follow because some are global and some are attached to bullets, yet they're all at the same level).

feel free to drop this commit if it doesn't work for the actual markdown parser.