test(queryCache): add failing test for partially undefined key (#3741)
The current behavior when using functions such as invalidateQueries or setQueriesData
is unexpected when objects with undefined properties are involved.

A key containing an object with an undefined property is hashed without the undefined property,
yet the property is considered in the partialDeepEqual function.

This creates some confusing scenarios, as demonstrated in the discussion on TkDodo's blog
<https://github.com/TkDodo/blog-comments/discussions/71#discussioncomment-4348406> and in the
referenced issue

This commit includes a failing test to demonstrate the expected behavior