chore: remove createStore option (#5352)
having the option to customize the internal store sound nice, but the problem is that the store holds Query instances, not data. So, the setter is only called once, not on every data update. This makes it relatively useless - every time we thought we could leverage createStore, it fell short of our requirements because of this.

The better way is usually to subscribe to the QueryCache directly