feat: add stable-query-client rule
This commit introduces a new rule in the eslint-plugin-query package. The rule, named 'stable-query-client', ensures that the QueryClient is
stable. It checks if the QueryClient is either extracted from the component or wrapped in React.useState. If not, it suggests a fix to wrap
the QueryClient instantiation in a React.useState call. This rule is recommended to be set as an error.