fix(codemod): remove the function overloads when the second argument is a function expression (#6191)
* fix(codemod): remove the function overloads when the second argument is a function expression

After a quick sync with @TkDodo, it turned out in this case it's safe to do the transformation, because the second argument is a function expression, so it means the first parameter must be the query key.

* fix(codemod): add the `useQuery` hook to the transformation list, because it was missing

* chore(codemod): rename the `parameters` constant to `functionArguments`

The naming was wrong from the beginning. The codemod does transformations with the arguments of the functions, not with their parameters. :)

* chore(codemod): fix the import order

The CI was failing because of it.

---------

Co-authored-by: Dominik Dorfmeister <office@dorfmeister.cc>