fix(core): mutation GC leak when reset, but not unsubscribed (#6509)
mutations were not gc'd correctly when being reset. GC was invoked on unsubscribe, but reset disconnects the observer from the mutation in the cache and must thus also trigger a GC, because the observer is no longer watching the values of that mutation, and another mutate() call will create a new mutation anyway. Thus, the old mutation becomes orphaned when reset is called.