fix(react): re-introduce setting options before suspending
if some options are set on the useSuspenseQuery observer instance, they won't be set in cases where the query fails and thus never re-mounts the observer. This is troublesome for e.g. the `retry` option because it means setting retries will not be respected. retry is needed when erroring, so it might be specific to that, but also, a lower `gcTime` would not not be set properly for example.