refactor(angular-query): change query functions to accept an options object instead of an injector
change query functions to accept an options object instead of an injector so that it matches angular APIs such as effect and allows adding more options in the future.

BREAKING CHANGE: The injector option has been replaced with an options object. The injector can now be provided with the options object.
Closes #8711