fix(query-core): fix queriesObserver result and observer length mismatch (#9973)
In 1b54538, `setOptions` was changed to call `notifyListeners` which
eventually leads to `QueriesObserver.setQueries`

1. Updating `#observerMatches`
2. Calling `setOptions` which leads to `#notify` being called
3. Updating `#result`

This is an issue when `#notify` is called before the `#result` is
updated, as `#trackResult` assumes these two are the same length, which
may not be the case if the number of queries changed. When the length
increases, `observerResult` becomes `undefined` and the function throws
an exception.

Fix by moving the `#observerMatches` assignment later in the function,
next to the `#observers` and `#result` assignment.

Co-authored-by: nrahnema <nrahnema@hudson-trading.com>
Co-authored-by: Dominik Dorfmeister <office@dorfmeister.cc>